pom/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
fastapi-mail==1.4.1


pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
google-generativeai>=0.3.0
opencv-python-headless==4.8.1.78
numpy==1.26.2

python-magic==0.4.27
